Feat: Added Player Routes and Controller
Fix: Added Id Format Checker

Chore: Add Logic for seasons and Matches

+++ backend/src/services/team-service.js
@@ -1,4 +1,5 @@
 const Team = require("../models/team-model");
+const mongoose = require("mongoose");
 
 class TeamService {
   async createTeam(teamData) {
@@ -20,13 +21,18 @@ class TeamService {
     return teamList;
   }
 
-  async getTeamById(temaId) {
-    let team = await Team.findById(temaId);
+  async getTeamById(teamId) {
+    let team = await Team.findById(teamId);
+    if (!mongoose.Types.ObjectId.isValid(teamId))
+      throw new Error("Invalid Id Format");
     if (!team) throw new Error("Team not found");
     return team;
   }
 
   async updateTeam(teamId, teamData) {
+    if (!teamId) throw new Error("TeamId is Required");
+    if (!mongoose.Types.ObjectId.isValid(teamId))
+      throw new Error("Invalid Id Format");
     const { name, logo } = teamData;
     if (!name || !logo) throw new Error("All fields are required");
     let updatedTeam = await Team.findByIdAndUpdate(
@@ -39,6 +45,9 @@ class TeamService {
   }
 
   async deleteTeam(teamId) {
+    if (!teamId) throw new Error("TeamId is Required");
+    if (!mongoose.Types.ObjectId.isValid(teamId))
+      throw new Error("Invalid Id Format");
     let deletedTeam = await Team.findByIdAndDelete(teamId);
     if (!deletedTeam) throw new Error("Error deleting team");
     return { msg: "Team deleted successfully" };
